# controllers/auth.controller.go
package controllers

import (
	"fmt"
	"github.com/gofiber/fiber/v2"
	"github.com/golang-jwt/jwt"
	"github.com/google/uuid"
	"golang.org/x/crypto/bcrypt"
	"tiktok-arena/configuration"
	"tiktok-arena/database"
	"tiktok-arena/models"
	"tiktok-arena/utils"
	"time"
)

//	@Summary		Register user
//	@Description	Register new user with given credentials
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			payload			body		models.RegisterInput	true	"Data to register user"
//	@Success		200				{object}	models.UserInfo			"Register success"
//	@Failure		400				{object}	models.FiberMessage		"Failed to register user"
//	@Router			/auth/register	[post]
func RegisterUser(c *fiber.Ctx) error {
	var payload *models.RegisterInput

	err := c.BodyParser(&payload)
	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, err.Error())
	}

	err = models.ValidateStruct(payload)
	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, err.Error())
	}

	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(payload.Password), bcrypt.DefaultCost)

	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, err.Error())
	}

	if database.CheckIfUserExists(payload.Name) {
		return utils.FiberMessage(c, fiber.StatusBadRequest,
			fmt.Sprintf("User %s already exists", payload.Name))
	}

	newUser := models.User{
		Name:     payload.Name,
		Password: string(hashedPassword),
	}

	err = database.CreateNewUser(&newUser)

	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadGateway, err.Error())
	}

	token, err := UserJwtToken(&newUser)

	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadGateway,
			fmt.Sprintf("Generating JWT Token failed: %v", err))
	}

	return c.Status(fiber.StatusCreated).JSON(
		models.AuthDetails{
			Username: newUser.Name,
			Token:    token,
		},
	)
}

//	@Summary		Login user
//	@Description	Login user with given credentials
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			payload			body		models.LoginInput	true	"Data to login user"
//	@Success		200				{object}	models.UserInfo		"Login success"
//	@Failure		400				{object}	models.FiberMessage	"Error logging in"
//	@Router			/auth/login    	[post]
func LoginUser(c *fiber.Ctx) error {
	var payload *models.LoginInput

	err := c.BodyParser(&payload)
	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, err.Error())
	}

	err = models.ValidateStruct(payload)
	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, err.Error())
	}

	user, err := database.GetUserByName(payload.Name)
	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, "Invalid credentials")
	}

	err = bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(payload.Password))
	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, "Invalid credentials")
	}

	token, err := UserJwtToken(&user)

	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadGateway,
			fmt.Sprintf("Generating JWT Token failed: %v", err))
	}

	return c.Status(fiber.StatusOK).JSON(
		models.AuthDetails{
			Username: user.Name,
			Token:    token,
		},
	)
}

func UserJwtToken(user *models.User) (string, error) {
	now := time.Now().UTC()

	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
		"sub":  user.ID,
		"name": user.Name,
		"exp":  now.Add(configuration.EnvConfig.JwtExpiresIn).Unix(),
		"iat":  now.Unix(),
		"nbf":  now.Unix(),
	})

	tokenString, err := token.SignedString([]byte(configuration.EnvConfig.JwtSecret))

	if err != nil {
		return "", err
	}

	return tokenString, nil
}

//	@Summary		Authenticated user details
//	@Description	Get current user id and name
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Success		200	{object}	models.UserInfo		"User details"
//	@Failure		400	{object}	models.FiberMessage	"Error getting user data"
//	@Router			/auth/whoami [get]
func WhoAmI(c *fiber.Ctx) error {
	user := c.Locals("user").(*jwt.Token)
	claims := user.Claims.(jwt.MapClaims)
	username := claims["name"].(string)
	id, err := uuid.Parse(claims["sub"].(string))

	if err != nil {
		return utils.FiberMessage(c, fiber.StatusBadRequest, "Error while parsing user id")
	}

	return c.Status(fiber.StatusOK).JSON(models.UserInfo{
		ID:   &id,
		Name: username,
	})
}